testcases/utility: crashing when "Nonetype" is returned
In hardware.py, the test for memory configured speed
doesn't check if the returned type is of None and still tries to split.
Which will cause the program to crash as you can't split None.

None type is returned if the user doesn't have a Configured Memory Speed
in their dmidecode. Now if None type is returned we send an "Unknown".